Use artificial circuits again

+++ src/lexer.rs
@@ -16,7 +16,7 @@ pub enum Symbol {
 /// The set of possible tokens.
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub enum Token {
-    Num(f64),
+    Num(i64),
     Sym(Symbol),
 }
 
@@ -132,7 +132,7 @@ pub fn lex(s: &str) -> LexResult {
                 num.push(chars.next().unwrap().1);
                 comma_last = false;
             }
-            tokens.add(Token::Num(num.parse::<f64>().unwrap()));
+            tokens.add(Token::Num(num.parse::<i64>().unwrap()));
             continue;
         }
